Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/compositing/2d.composite.operation.darker.worker-expected.txt


FAIL 2d assert_equals: ctx.globalCompositeOperation === 'xor' (got darker[string], expected xor[string]) expected "xor" but got "darker"
